use double instead of float for the position field
This has been chosen because of precision issues in case
the represented joint is a wheel. Double gives ample range
while float is really too limited.

See http://rock.opendfki.de/wiki/WikiStart/OngoingWork/BaseTypesCleanup/JointState

+++ base/JointState.hpp
@@ -34,7 +34,7 @@ namespace base
          * If the joint is an unconstrained angular joint (e.g. a wheel joint),
          * the range is [-inf, inf]
          */
-        float position;
+        double position;
 
         /** Speed in radians per second for angular actuators, in m/s
          * for linear ones
